LangGraph_Course/last_project/requirements.txt
langgraph
streamlit
langchain
python-dotenv
deepinfra
tavily-python
